requirements.txt
aiogram==3.13.0
apscheduler==3.10.4
openai==1.51.0
anthropic==0.40.0
python-dotenv==1.0.1
requests==2.32.3
